Fix a majors warnings

+++ GolemEditor/Source/UI/Windows/viewport.cpp
@@ -100,7 +100,7 @@ void Viewport::Update()
         m_camera->ProcessKeyboardInput(GolemEngine::GetDeltaTime());
         ImGui::SetMouseCursor(ImGuiMouseCursor_None);
         // Update camera with mouse position
-        m_camera->ProcessMouseMovement(InputManager::GetMouseWindowPos(), true, windowDimensions, ImGui::GetMousePos().x, ImGui::GetMousePos().y);
+        m_camera->ProcessMouseMovement(InputManager::GetMouseWindowPos(), true, windowDimensions, (int)ImGui::GetMousePos().x, (int)ImGui::GetMousePos().y);
         // Update camera speed depending on scroll
         m_camera->ProcessMouseScroll(InputManager::GetMouseScroll());
         InputManager::SetMouseScroll(0.0f);     // Otherwise the camera will continue to change since GetMouseScroll value doesn't change but has a value
